Move remaining files from "output" package to "cache"
