--- gluecodium/src/main/java/com/here/gluecodium/output/package-info.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (C) 2016-2019 HERE Europe B.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- * License-Filename: LICENSE
- */
-
-/**
- * Simple redirection of output to file or console.
- *
- * <h1>Input</h1>
- *
- * List of generated files.
- *
- * <h1>Output</h1>
- *
- * Redirection to file or console output.
- */
-package com.here.gluecodium.output;
